Rename package 'apiutil' to 'api'
