+++ api/endpoints.go
@@ -1,4 +1,4 @@
-package apiutil
+package api
 
 import (
 	"encoding/json"
