+++ api/endpoints_test.go
@@ -1,4 +1,4 @@
-package apiutil
+package api
 
 import (
 	"net/http/httptest"
